Move InMemoryRepository to Repositories::InMemory
