+++ ruby_event_store/lib/ruby_event_store.rb
@@ -1,6 +1,6 @@
 require 'ruby_event_store/pub_sub/dispatcher'
 require 'ruby_event_store/pub_sub/broker'
-require 'ruby_event_store/in_memory_repository'
+require 'ruby_event_store/repositories/in_memory'
 require 'ruby_event_store/projection'
 require 'ruby_event_store/errors'
 require 'ruby_event_store/constants'
