+++ ruby_event_store/spec/client_spec.rb
@@ -6,25 +6,25 @@ module RubyEventStore
     TestEvent = Class.new(RubyEventStore::Event)
 
     specify 'publish_event returns :ok when success' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       expect(client.publish_event(TestEvent.new)).to eq(:ok)
     end
 
     specify 'append_to_stream returns :ok when success' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       expect(client.append_to_stream(TestEvent.new, stream_name: stream)).to eq(:ok)
     end
 
     specify 'append to default stream when not specified' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       test_event = TestEvent.new
       expect(client.append_to_stream(test_event)).to eq(:ok)
       expect(client.read_stream_events_forward(GLOBAL_STREAM)).to eq([test_event])
     end
 
     specify 'publish to default stream when not specified' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       test_event = TestEvent.new
       expect(client.publish_events([test_event])).to eq(:ok)
       expect(client.read_stream_events_forward(GLOBAL_STREAM)).to eq([test_event])
@@ -32,12 +32,12 @@ module RubyEventStore
 
     specify 'delete_stream returns :ok when success' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       expect(client.delete_stream(stream)).to eq(:ok)
     end
 
     specify 'publish to default stream when not specified' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       test_event = TestEvent.new
       expect(client.publish_event(test_event)).to eq(:ok)
       expect(client.read_stream_events_forward(GLOBAL_STREAM)).to eq([test_event])
@@ -45,7 +45,7 @@ module RubyEventStore
 
     specify 'publish first event, expect any stream state' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       expect(client.publish_event(first_event, stream_name: stream)).to eq(:ok)
       expect(client.read_stream_events_forward(stream)).to eq([first_event])
@@ -53,7 +53,7 @@ module RubyEventStore
 
     specify 'publish next event, expect any stream state' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       client.append_to_stream(first_event, stream_name: stream)
@@ -63,7 +63,7 @@ module RubyEventStore
 
     specify 'publish first event, expect empty stream' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       expect(client.publish_event(first_event, stream_name: stream, expected_version: :none)).to eq(:ok)
       expect(client.read_stream_events_forward(stream)).to eq([first_event])
@@ -71,7 +71,7 @@ module RubyEventStore
 
     specify 'publish first event, fail if not empty stream' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       client.append_to_stream(first_event, stream_name: stream)
@@ -81,7 +81,7 @@ module RubyEventStore
 
     specify 'publish event, expect last event to be the last read one' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       client.append_to_stream(first_event, stream_name: stream)
@@ -91,7 +91,7 @@ module RubyEventStore
 
     specify 'publish event, fail if last event is not the last read one' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       third_event   = TestEvent.new
@@ -103,7 +103,7 @@ module RubyEventStore
 
     specify 'append many events' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       client.append_to_stream([first_event, second_event], stream_name: stream, expected_version: -1)
@@ -112,7 +112,7 @@ module RubyEventStore
 
     specify 'read only up to page size from stream' do
       stream = SecureRandom.uuid
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       (1..102).each { client.append_to_stream(TestEvent.new, stream_name: stream) }
       expect(client.read_events_forward(stream, count: 10).size).to eq(10)
       expect(client.read_events_backward(stream, count: 10).size).to eq(10)
@@ -126,7 +126,7 @@ module RubyEventStore
     end
 
     specify 'published event metadata will be enriched by proc execution' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new, metadata_proc: ->{ {request_id: '127.0.0.1'} })
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new, metadata_proc: ->{ {request_id: '127.0.0.1'} })
       event = TestEvent.new
       client.publish_event(event)
       published = client.read_all_streams_forward
@@ -136,7 +136,7 @@ module RubyEventStore
     end
 
     specify 'only timestamp set inn metadata when event stored in stream if metadata proc return nil' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new, metadata_proc: ->{ nil })
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new, metadata_proc: ->{ nil })
       event = TestEvent.new
       client.append_to_stream(event)
       published = client.read_all_streams_forward
@@ -150,7 +150,7 @@ module RubyEventStore
       utc = Time.parse('2015-05-04 13:17:23 UTC')
       allow(Time).to receive(:now).and_return(now)
       allow_any_instance_of(Time).to receive(:utc).and_return(utc)
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       event = TestEvent.new
       client.publish_event(event)
       published = client.read_all_streams_forward
@@ -159,20 +159,20 @@ module RubyEventStore
     end
 
     specify 'throws exception if subscriber is not defined' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       expect { client.subscribe(nil, [])}.to raise_error(SubscriberNotExist)
       expect { client.subscribe_to_all_events(nil)}.to raise_error(SubscriberNotExist)
     end
 
     specify 'reading all existing stream names' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       client.publish_event(TestEvent.new, stream_name: 'test')
 
       expect(client.get_all_streams).to eq([Stream.new('all'), Stream.new('test')])
     end
 
     specify 'reading particular event' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       test_event = TestEvent.new
       client.publish_event(test_event, stream_name: 'test')
 
@@ -180,12 +180,12 @@ module RubyEventStore
     end
 
     specify 'reading non-existent event' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       expect{client.read_event('72922e65-1b32-4e97-8023-03ae81dd3a27')}.to raise_error(EventNotFound)
     end
 
     specify 'link events' do
-      client = RubyEventStore::Client.new(repository: InMemoryRepository.new)
+      client = RubyEventStore::Client.new(repository: Repositories::InMemory.new)
       first_event   = TestEvent.new
       second_event  = TestEvent.new
       client.subscribe_to_all_events(subscriber = Subscribers::ValidHandler.new)
